ResourceGroupsTaggingAPI: WorkspacesWeb: Fail gracefully for unsupported regions
